researches/MAI-UI/mai_ui_dev/requirements_cloudphone.txt
# MAI-UI Cloud Phone Web UI 依赖

# Web 框架
gradio>=4.0.0

# 图像处理
Pillow>=10.0.0

# 配置文件
pyyaml>=6.0

# HTTP 请求
requests>=2.28.0
